refactor: split cursor api and cache helpers
