+++ docs/CODEBASE_AUDIT_ACTION_TABLE_2026-03-09.md
@@ -43,6 +43,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | R23 | Fixed | Telemetry query-layer split | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_queries.go`, `internal/telemetry/usage_view_languages.go` | The SQL aggregation/query helpers and language-inference helpers moved out of the main usage-view file into dedicated query/language units, leaving the main file focused on orchestration and shared aggregate types. | Continue shrinking orchestration/materialization into smaller units if needed. |
 | R24 | Fixed | OpenRouter generation-path split | `internal/providers/openrouter/openrouter.go`, `internal/providers/openrouter/generations.go` | Generation payload types and generation-fetch/enrichment/aggregation logic moved out of the main OpenRouter provider file into a dedicated generation unit. | Continue with analytics/client/API helper splits. |
 | R25 | Fixed | OpenRouter clock injection | `internal/providers/openrouter/openrouter.go`, `internal/providers/openrouter/generations.go` | OpenRouter’s time-sensitive analytics and generation flows now use an injectable clock instead of reading `time.Now()` directly in the provider hot path. | Extend the same pattern to remaining providers and analytics helpers. |
+| R26 | Fixed | Cursor API/cache helper split | `internal/providers/cursor/cursor.go`, `internal/providers/cursor/api.go`, `internal/providers/cursor/cache.go` | Cursor HTTP helper methods and billing/model cache helpers now live in dedicated units instead of the main provider file, removing duplicate implementations and narrowing `cursor.go` to orchestration and local-data projection. | Continue splitting SQLite projection and token-loading responsibilities. |
 
 ## Action Table
 
@@ -52,7 +53,7 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | A2 | P2 | TUI/application decomposition follow-through | `internal/tui/model.go:393-584`, `internal/dashboardapp/service.go` | The side effects are now injected, but `Model` still owns a very large amount of event-handling and state-transition logic. | Continue splitting update/action logic into smaller TUI units and move more orchestration decisions into the dashboard application layer over time. | Lower UI complexity and smaller blast radius per change. |
 | A3 | P2 | UI metric-prefix parsing | `internal/tui/analytics.go:663-729`, `internal/tui/detail.go`, `internal/core/usage_breakdowns.go` | The main composition bars now consume shared extractors, but analytics/detail-specific sections still decode some raw metric-key conventions directly in TUI code. | Continue promoting the remaining analytics/detail extractors into `internal/core` or `internal/telemetry` so renderers consume structured sections instead of re-parsing maps. | Removes a large class of UI drift bugs and reduces per-render work. |
 | A4 | P1 | OpenRouter provider size | `internal/providers/openrouter/openrouter.go`, `internal/providers/openrouter/provider_resolution.go`, `internal/providers/openrouter/generations.go` | `openrouter.go` is materially smaller after the provider-resolution and generation-path splits, but it still mixes auth probing, credits, keys, analytics parsing, and some output/projection helpers. | Continue splitting into `api_client`, `analytics`, and remaining projection/helper units. | Easier maintenance, smaller diff surface, faster targeted testing. |
-| A5 | P1 | Cursor provider responsibility overload | `internal/providers/cursor/cursor.go:181-335`, `internal/providers/cursor/cursor.go:903-1006`, `internal/providers/cursor/cursor.go:1087-2086` | Cursor provider combines API orchestration, local SQLite readers, token extraction, and two independent caches in one class. | Split into `api`, `trackingdb`, `statedb`, `cache`, and `snapshot_projection` modules. Move token extraction out of provider hot path. | Cleaner boundaries and less risk of local/API logic regressions. |
+| A5 | P1 | Cursor provider responsibility overload | `internal/providers/cursor/cursor.go:181-335`, `internal/providers/cursor/cursor.go:903-1842`, `internal/providers/cursor/state_records.go`, `internal/providers/cursor/tracking_records.go` | Cursor provider no longer owns the API/cache helpers, but it still combines fetch orchestration, token loading, local SQLite projection, and some remaining snapshot assembly in one large file. | Continue splitting token-loading, local DB projection, and snapshot assembly into dedicated modules so `cursor.go` becomes a thin coordinator. | Cleaner boundaries and less risk of local/API logic regressions. |
 | A6 | P2 | Telemetry usage-view monolith | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_helpers.go`, `internal/telemetry/usage_view_projection.go`, `internal/telemetry/usage_view_queries.go` | The usage-view code is materially smaller after the helper/projection/query splits, but the orchestration/materialization path still owns temp-table lifecycle, query fanout, and aggregate assembly in one place. | Continue splitting remaining orchestration/materialization concerns and consider a typed intermediate aggregation model. | Easier optimization and safer incremental changes. |
 | A7 | P2 | Daemon service monolith | `internal/daemon/server.go`, `internal/daemon/server_http.go`, `internal/daemon/server_read_model.go`, `internal/daemon/server_loops.go` | The daemon is materially less coupled after the logging/cache/http/loop split, but polling, collection, retention, and spool maintenance still share one large runtime helper unit. | Continue splitting the loop-heavy runtime into `polling`, `collection`, `retention`, and `spool` units. | Lower mental load and easier concurrency review. |
 | A11 | P2 | Time-dependent logic without injectable clock | `internal/providers/ollama/ollama.go:1088`, `internal/core/analytics_normalize.go:61-103` | Cursor and OpenRouter now use injectable clocks in their main time-sensitive paths, but other providers and analytics helpers still read `time.Now()` directly, often mixing local time and UTC. | Extend the clock abstraction to the remaining provider and analytics subsystems and standardize UTC/local semantics per provider. | Better determinism and fewer timezone edge cases. |
