+++ internal/providers/cursor/cursor.go
@@ -1,15 +1,11 @@
 package cursor
 
 import (
-	"bytes"
 	"context"
 	"database/sql"
-	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"math"
-	"net/http"
 	"sort"
 	"strconv"
 	"strings"
@@ -701,81 +697,6 @@ func (p *Provider) fetchFromAPI(ctx context.Context, token string, snap *core.Us
 	return nil
 }
 
-func (p *Provider) callDashboardAPI(ctx context.Context, token, method string, result interface{}) error {
-	url := fmt.Sprintf("%s/aiserver.v1.DashboardService/%s", cursorAPIBase, method)
-	return p.doPost(ctx, token, url, result)
-}
-
-func (p *Provider) callDashboardAPIWithBody(ctx context.Context, token, method string, body []byte, result interface{}) error {
-	url := fmt.Sprintf("%s/aiserver.v1.DashboardService/%s", cursorAPIBase, method)
-	req, err := http.NewRequestWithContext(ctx, "POST", url, bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := p.Client().Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		respBody, _ := io.ReadAll(resp.Body)
-		return fmt.Errorf("HTTP %d: %s", resp.StatusCode, string(respBody))
-	}
-
-	return json.NewDecoder(resp.Body).Decode(result)
-}
-
-func (p *Provider) callRESTAPI(ctx context.Context, token, path string, result interface{}) error {
-	url := cursorAPIBase + path
-
-	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := p.Client().Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		body, _ := io.ReadAll(resp.Body)
-		return fmt.Errorf("HTTP %d: %s", resp.StatusCode, string(body))
-	}
-
-	return json.NewDecoder(resp.Body).Decode(result)
-}
-
-func (p *Provider) doPost(ctx context.Context, token, url string, result interface{}) error {
-	body := []byte("{}")
-	req, err := http.NewRequestWithContext(ctx, "POST", url, bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := p.Client().Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		respBody, _ := io.ReadAll(resp.Body)
-		return fmt.Errorf("HTTP %d: %s", resp.StatusCode, string(respBody))
-	}
-
-	return json.NewDecoder(resp.Body).Decode(result)
-}
-
 func applyModelAggregations(snap *core.UsageSnapshot, aggregations []modelAggregation) bool {
 	if len(aggregations) == 0 {
 		return false
@@ -924,190 +845,6 @@ func parseModelTokenCount(raw string) (float64, bool) {
 		return 0, false
 	}
 	return v, true
-}
-
-func (p *Provider) storeModelAggregationCache(accountID, billingCycleStart, billingCycleEnd string, aggregations []modelAggregation, effectiveLimitUSD float64) {
-	if accountID == "" || len(aggregations) == 0 {
-		return
-	}
-	copied := make([]modelAggregation, len(aggregations))
-	copy(copied, aggregations)
-
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	if p.modelAggregationCache == nil {
-		p.modelAggregationCache = make(map[string]cachedModelAggregation)
-	}
-	entry := cachedModelAggregation{
-		BillingCycleStart: billingCycleStart,
-		BillingCycleEnd:   billingCycleEnd,
-		Aggregations:      copied,
-		EffectiveLimitUSD: effectiveLimitUSD,
-	}
-	// Preserve billing metrics from a previous storeBillingMetricsCache call.
-	if prev, ok := p.modelAggregationCache[accountID]; ok && len(prev.BillingMetrics) > 0 {
-		entry.BillingMetrics = prev.BillingMetrics
-	}
-	p.modelAggregationCache[accountID] = entry
-}
-
-func (p *Provider) applyCachedModelAggregations(accountID, billingCycleStart, billingCycleEnd string, snap *core.UsageSnapshot) bool {
-	if accountID == "" {
-		return false
-	}
-
-	p.mu.RLock()
-	cached, ok := p.modelAggregationCache[accountID]
-	p.mu.RUnlock()
-	if !ok || len(cached.Aggregations) == 0 {
-		return false
-	}
-
-	if billingCycleStart != "" && cached.BillingCycleStart != "" && billingCycleStart != cached.BillingCycleStart {
-		return false
-	}
-	if billingCycleEnd != "" && cached.BillingCycleEnd != "" && billingCycleEnd != cached.BillingCycleEnd {
-		return false
-	}
-
-	copied := make([]modelAggregation, len(cached.Aggregations))
-	copy(copied, cached.Aggregations)
-	return applyModelAggregations(snap, copied)
-}
-
-// billingMetricKeys lists the metric keys cached for local-only fallback.
-var billingMetricKeys = []string{
-	"plan_spend", "plan_percent_used", "plan_auto_percent_used", "plan_api_percent_used",
-	"spend_limit", "individual_spend", "team_budget", "team_budget_self", "team_budget_others",
-	"plan_included", "plan_bonus", "plan_total_spend_usd", "plan_limit_usd",
-}
-
-func cloneMetric(m core.Metric) core.Metric {
-	out := core.Metric{Unit: m.Unit, Window: m.Window}
-	if m.Limit != nil {
-		out.Limit = core.Float64Ptr(*m.Limit)
-	}
-	if m.Remaining != nil {
-		out.Remaining = core.Float64Ptr(*m.Remaining)
-	}
-	if m.Used != nil {
-		out.Used = core.Float64Ptr(*m.Used)
-	}
-	return out
-}
-
-// storeBillingMetricsCache snapshots the current billing metrics so they can
-// be restored when the API is temporarily unavailable.
-func (p *Provider) storeBillingMetricsCache(accountID string, snap *core.UsageSnapshot) {
-	if accountID == "" {
-		return
-	}
-	cached := make(map[string]core.Metric, len(billingMetricKeys))
-	for _, key := range billingMetricKeys {
-		if m, ok := snap.Metrics[key]; ok {
-			cached[key] = cloneMetric(m)
-		}
-	}
-	if len(cached) == 0 {
-		return
-	}
-
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	if p.modelAggregationCache == nil {
-		p.modelAggregationCache = make(map[string]cachedModelAggregation)
-	}
-	entry := p.modelAggregationCache[accountID]
-	entry.BillingMetrics = cached
-	p.modelAggregationCache[accountID] = entry
-}
-
-// applyCachedBillingMetrics restores billing metrics from cache into the
-// snapshot so that credit gauges render when the API is temporarily down.
-func (p *Provider) applyCachedBillingMetrics(accountID string, snap *core.UsageSnapshot) {
-	if accountID == "" {
-		return
-	}
-	p.mu.RLock()
-	cached, ok := p.modelAggregationCache[accountID]
-	p.mu.RUnlock()
-	if !ok || len(cached.BillingMetrics) == 0 {
-		return
-	}
-	for key, m := range cached.BillingMetrics {
-		if _, exists := snap.Metrics[key]; !exists {
-			snap.Metrics[key] = cloneMetric(m)
-		}
-	}
-}
-
-// ensureCreditGauges synthesizes credit metrics from local data when API
-// didn't provide them. This runs as a final step in Fetch() so the Credits
-// tag and gauge bars render regardless of API availability.
-func (p *Provider) ensureCreditGauges(accountID string, snap *core.UsageSnapshot) {
-	// Already have gauge-eligible credit metrics from API — nothing to do.
-	if _, ok := snap.Metrics["plan_spend"]; ok {
-		return
-	}
-	if _, ok := snap.Metrics["spend_limit"]; ok {
-		return
-	}
-
-	// Determine total cost from best available source.
-	var costUSD float64
-	if m, ok := snap.Metrics["billing_total_cost"]; ok && m.Used != nil && *m.Used > 0 {
-		costUSD = *m.Used
-	} else if m, ok := snap.Metrics["composer_cost"]; ok && m.Used != nil && *m.Used > 0 {
-		costUSD = *m.Used
-	}
-	if costUSD <= 0 {
-		return
-	}
-
-	// Always expose plan_total_spend_usd so the Credits tag renders in the
-	// TUI even without a limit (computeDisplayInfoRaw checks this key).
-	if _, ok := snap.Metrics["plan_total_spend_usd"]; !ok {
-		snap.Metrics["plan_total_spend_usd"] = core.Metric{
-			Used:   core.Float64Ptr(costUSD),
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-	}
-
-	// Try to find a limit so we can create a gauge bar.
-	var limitUSD float64
-
-	// 1) From plan_included_amount (GetPlanInfo may have succeeded).
-	if m, ok := snap.Metrics["plan_included_amount"]; ok && m.Used != nil && *m.Used > 0 {
-		limitUSD = *m.Used
-	}
-
-	// 2) From cached effective limit.
-	if limitUSD <= 0 {
-		p.mu.RLock()
-		if cached, ok := p.modelAggregationCache[accountID]; ok && cached.EffectiveLimitUSD > 0 {
-			limitUSD = cached.EffectiveLimitUSD
-		}
-		p.mu.RUnlock()
-	}
-
-	// 3) From plan_included_amount_cents in Raw (may have been set by API).
-	if limitUSD <= 0 {
-		if raw, ok := snap.Raw["plan_included_amount_cents"]; ok {
-			if cents, err := strconv.ParseFloat(raw, 64); err == nil && cents > 0 {
-				limitUSD = cents / 100.0
-			}
-		}
-	}
-
-	if limitUSD > 0 {
-		snap.Metrics["plan_spend"] = core.Metric{
-			Used:   core.Float64Ptr(costUSD),
-			Limit:  core.Float64Ptr(limitUSD),
-			Unit:   "USD",
-			Window: "billing-cycle",
-		}
-	}
 }
 
 func (p *Provider) readTrackingDB(ctx context.Context, dbPath string, snap *core.UsageSnapshot) error {
